Move ORM request to menu file
